Reorganise stats screen
The stats screen was changed to fit all the data on a single HD page. This was great, but the ordering, due to being the same as the SD layout, didn't really make sense. For example, Disarmed by was positioned somewhere in the middle of the stats. This PR reorders the stats. They should make more sense on both SD and HD now.

- Increased min voltage start value. This will be good up to 26S now. Yes, I know someone who is looking at a build with that much voltage.
- Added stats for min/max sats
- Separated out the stats draw for each type of stat in to its own function. This makes the displaying simpler. Plus it will be ready for customisable stats screen if someone does that.
- There is room for 3 more stats on WTF and 5 on analogue.